Fix namespacing of all files but add shim redirects in
